Adding cross-platform supports for windows users

+++ qqtt/engine/trainer_warp.py
@@ -1435,7 +1435,7 @@ def get_simple_shadow(
     shadow_image[valid_pixel_coords[:, 1], valid_pixel_coords[:, 0]] = 255
 
     kernel = np.ones((kernel_size, kernel_size), np.uint8)
-    kernel_1 = np.ones((3, 3), np.uint(8))
+    kernel_1 = np.ones((3, 3), np.uint8(8))
     dilated_shadow = cv2.dilate(shadow_image, kernel, iterations=1)
     dilated_shadow = cv2.dilate(dilated_shadow, kernel_1, iterations=1)
     final_shadow = cv2.erode(dilated_shadow, kernel, iterations=1)
